--- a/galois/_codes/_bch.py
+++ b/galois/_codes/_bch.py
@@ -11,11 +11,12 @@ import numba
 from numba import int64
 import numpy as np
 
-from .._domains._function import JITFunction
+from .._domains._function import Function
 from .._fields import Field, FieldArray, GF2  # pylint: disable=unused-import
 from .._lfsr import berlekamp_massey_jit
 from .._overrides import set_module
 from .._polys import Poly, matlab_primitive_poly
+from .._polys._dense import roots_jit, divmod_jit
 from .._prime import factors
 from ..typing import PolyLike
 
@@ -775,12 +776,12 @@ class BCH:
         syndrome = codeword.view(self.field) @ self.H[:,-ns:].T
 
         # Invoke the JIT compiled function
-        dec_codeword, N_errors = decode_jit.call(self.field, codeword, syndrome, self.t, int(self.field.primitive_element))
+        dec_codeword, N_errors = decode_jit(self.field)(codeword, syndrome, self.t, int(self.field.primitive_element))
 
         if self.systematic:
             message = dec_codeword[:, 0:ks]
         else:
-            message, _ = GF2._poly_divmod(dec_codeword[:, 0:ns].view(GF2), self.generator_poly.coeffs)
+            message, _ = divmod_jit(GF2)(dec_codeword[:, 0:ns].view(GF2), self.generator_poly.coeffs)
         message = message.view(type(codeword))  # TODO: Remove this
 
         if codeword_1d:
@@ -970,7 +971,7 @@ class BCH:
         return self._is_narrow_sense
 
 
-class decode_jit(JITFunction):
+class decode_jit(Function):
     """
     Performs BCH decoding.
 
@@ -978,31 +979,23 @@ class decode_jit(JITFunction):
     ----------
     * Lin, S. and Costello, D. Error Control Coding. Section 7.4.
     """
-    _CACHE = {}
-
-    @classmethod
-    def call(cls, field, codeword, syndrome, t, primitive_element):
-        if field.ufunc_mode != "python-calculate":
-            codeword_ = codeword.astype(np.int64)
-            syndrome_ = syndrome.astype(np.int64)
-            y = cls.jit(field)(codeword_, syndrome_, t, primitive_element)
+    def __call__(self, codeword, syndrome, t, primitive_element):
+        if self.field.ufunc_mode != "python-calculate":
+            y = self.jit(codeword.astype(np.int64), syndrome.astype(np.int64), t, primitive_element)
         else:
-            codeword_ = codeword.view(np.ndarray)
-            syndrome_ = syndrome.view(np.ndarray)
-            y = cls.python(field)(codeword_, syndrome_, t, primitive_element)
+            y = self.python(codeword.view(np.ndarray), syndrome.view(np.ndarray), t, primitive_element)
 
         dec_codeword, N_errors = y[:,0:-1], y[:,-1]
         dec_codeword = dec_codeword.astype(codeword.dtype)
-        dec_codeword = dec_codeword.view(field)
+        dec_codeword = dec_codeword.view(self.field)
 
         return dec_codeword, N_errors
 
-    @classmethod
-    def set_globals(cls, field: Type[FieldArray]):
+    def set_globals(self):
         # pylint: disable=global-variable-undefined
         global POLY_ROOTS, BERLEKAMP_MASSEY
-        POLY_ROOTS = field._function("poly_roots")
-        BERLEKAMP_MASSEY = berlekamp_massey_jit.function(field)
+        POLY_ROOTS = roots_jit(self.field).function
+        BERLEKAMP_MASSEY = berlekamp_massey_jit(self.field).function
 
     _SIGNATURE = numba.types.FunctionType(int64[:,:](int64[:,:], int64[:,:], int64, int64))