… benchmarks/test_field_arithmetic.py
… docs/performance/benchmarks.rst
… docs/requirements.txt
… galois/_domains/_arithmetic.py
… galois/_domains/_calculate.py
… galois/_domains/_function.py
… galois/_domains/_linalg.py
… galois/_domains/_lookup.py
… galois/_domains/_meta.py
… galois/_domains/_poly.py
… galois/_domains/_ufunc.py
… galois/_fields/_array.py
… galois/_fields/_factory.py
… galois/_fields/_gf2.py
… galois/_fields/_ufunc.py
… galois/_fields/_ufunc_2_1.py
… galois/_fields/_ufunc_2_m.py
… galois/_fields/_ufunc_p_1.py
… galois/_fields/_ufunc_p_m.py
… galois/_lfsr.py
… galois/_ntt.py
… galois/_polys/_binary.py
… galois/_polys/_dense.py
… galois/_polys/_functions.py
… galois/_polys/_poly.py
… galois/_polys/_primitive.py
… galois/_polys/_sparse.py
… tests/fields/test_arithmetic.py
… tests/fields/test_fft.py